update complex forms to use the new adapter pattern
