+++ backend/FwLite/LcmCrdt.Tests/JsonPatchEntryRewriteTests.cs
@@ -1,9 +1,7 @@
 ﻿using LcmCrdt.Changes.Entries;
 using LcmCrdt.Objects;
-using MiniLcm.Models;
 using SIL.Harmony.Changes;
 using SystemTextJsonPatch;
-using Entry = LcmCrdt.Objects.Entry;
 
 namespace LcmCrdt.Tests;
 
@@ -17,7 +15,7 @@ public class JsonPatchEntryRewriteTests
         var patch = new JsonPatchDocument<MiniLcm.Models.Entry>();
         var componentEntry = new Entry() { Id = Guid.NewGuid(), LexemeForm = { { "en", "component" } } };
         patch.Add(entry => entry.Components, ComplexFormComponent.FromEntries(_entry, componentEntry));
-        var changes = Entry.ChangesFromJsonPatch(_entry, patch);
+        var changes = _entry.ToChanges(patch);
         var addEntryComponentChange =
             changes.Should().ContainSingle().Which.Should().BeOfType<AddEntryComponentChange>().Subject;
         addEntryComponentChange.ComplexFormEntryId.Should().Be(_entry.Id);
@@ -33,9 +31,9 @@ public class JsonPatchEntryRewriteTests
         var complexFormComponent = ComplexFormComponent.FromEntries(_entry, componentEntry);
         _entry.Components.Add(complexFormComponent);
         patch.Remove(entry => entry.Components, 0);
-        var changes = Entry.ChangesFromJsonPatch(_entry, patch);
+        var changes = _entry.ToChanges(patch);
         var removeEntryComponentChange = changes.Should().ContainSingle().Which.Should()
-            .BeOfType<DeleteChange<CrdtComplexFormComponent>>().Subject;
+            .BeOfType<DeleteChange<ComplexFormComponent>>().Subject;
         removeEntryComponentChange.EntityId.Should().Be(complexFormComponent.Id);
     }
 
@@ -48,7 +46,7 @@ public class JsonPatchEntryRewriteTests
         _entry.Components.Add(complexFormComponent);
         var newComponentId = Guid.NewGuid();
         patch.Replace(entry => entry.Components, complexFormComponent with { ComponentEntryId = newComponentId, ComponentHeadword = "new" }, 0);
-        var changes = Entry.ChangesFromJsonPatch(_entry, patch);
+        var changes = _entry.ToChanges(patch);
         var setComplexFormComponentChange = changes.Should().ContainSingle().Which.Should()
             .BeOfType<SetComplexFormComponentChange>().Subject;
         setComplexFormComponentChange.EntityId.Should().Be(complexFormComponent.Id);
@@ -66,7 +64,7 @@ public class JsonPatchEntryRewriteTests
         _entry.Components.Add(complexFormComponent);
         var newComplexFormId = Guid.NewGuid();
         patch.Replace(entry => entry.Components, complexFormComponent with { ComplexFormEntryId = newComplexFormId, ComplexFormHeadword = "new" }, 0);
-        var changes = Entry.ChangesFromJsonPatch(_entry, patch);
+        var changes = _entry.ToChanges(patch);
         var setComplexFormComponentChange = changes.Should().ContainSingle().Which.Should()
             .BeOfType<SetComplexFormComponentChange>().Subject;
         setComplexFormComponentChange.EntityId.Should().Be(complexFormComponent.Id);
@@ -81,7 +79,7 @@ public class JsonPatchEntryRewriteTests
         var patch = new JsonPatchDocument<MiniLcm.Models.Entry>();
         var componentEntry = new Entry() { Id = Guid.NewGuid(), LexemeForm = { { "en", "complex form" } } };
         patch.Add(entry => entry.ComplexForms, ComplexFormComponent.FromEntries(_entry, componentEntry));
-        var changes = Entry.ChangesFromJsonPatch(componentEntry, patch);
+        var changes = componentEntry.ToChanges(patch);
         var addEntryComponentChange =
             changes.Should().ContainSingle().Which.Should().BeOfType<AddEntryComponentChange>().Subject;
         addEntryComponentChange.ComplexFormEntryId.Should().Be(_entry.Id);
@@ -97,9 +95,9 @@ public class JsonPatchEntryRewriteTests
         var complexFormComponent = ComplexFormComponent.FromEntries(_entry, componentEntry);
         componentEntry.ComplexForms.Add(complexFormComponent);
         patch.Remove(entry => entry.ComplexForms, 0);
-        var changes = Entry.ChangesFromJsonPatch(componentEntry, patch);
+        var changes = componentEntry.ToChanges(patch);
         var removeEntryComponentChange = changes.Should().ContainSingle().Which.Should()
-            .BeOfType<DeleteChange<CrdtComplexFormComponent>>().Subject;
+            .BeOfType<DeleteChange<ComplexFormComponent>>().Subject;
         removeEntryComponentChange.EntityId.Should().Be(complexFormComponent.Id);
     }
 
@@ -112,7 +110,7 @@ public class JsonPatchEntryRewriteTests
         componentEntry.ComplexForms.Add(complexFormComponent);
         var newComponentId = Guid.NewGuid();
         patch.Replace(entry => entry.ComplexForms, complexFormComponent with { ComponentEntryId = newComponentId, ComponentHeadword = "new" }, 0);
-        var changes = Entry.ChangesFromJsonPatch(componentEntry, patch);
+        var changes = componentEntry.ToChanges(patch);
         var setComplexFormComponentChange = changes.Should().ContainSingle().Which.Should()
             .BeOfType<SetComplexFormComponentChange>().Subject;
         setComplexFormComponentChange.EntityId.Should().Be(complexFormComponent.Id);
@@ -130,7 +128,7 @@ public class JsonPatchEntryRewriteTests
         componentEntry.ComplexForms.Add(complexFormComponent);
         var newComplexFormId = Guid.NewGuid();
         patch.Replace(entry => entry.ComplexForms, complexFormComponent with { ComplexFormEntryId = newComplexFormId, ComplexFormHeadword = "new" }, 0);
-        var changes = Entry.ChangesFromJsonPatch(componentEntry, patch);
+        var changes = componentEntry.ToChanges(patch);
         var setComplexFormComponentChange = changes.Should().ContainSingle().Which.Should()
             .BeOfType<SetComplexFormComponentChange>().Subject;
         setComplexFormComponentChange.EntityId.Should().Be(complexFormComponent.Id);
@@ -145,7 +143,7 @@ public class JsonPatchEntryRewriteTests
         var patch = new JsonPatchDocument<MiniLcm.Models.Entry>();
         var complexFormType = new ComplexFormType() { Id = Guid.NewGuid(), Name = new MultiString() };
         patch.Add(entry => entry.ComplexFormTypes, complexFormType);
-        var changes = Entry.ChangesFromJsonPatch(_entry, patch);
+        var changes = _entry.ToChanges(patch);
         var addComplexFormTypeChange =
             changes.Should().ContainSingle().Which.Should().BeOfType<AddComplexFormTypeChange>().Subject;
         addComplexFormTypeChange.EntityId.Should().Be(_entry.Id);
@@ -159,7 +157,7 @@ public class JsonPatchEntryRewriteTests
         var complexFormType = new ComplexFormType() { Id = Guid.NewGuid(), Name = new MultiString() };
         _entry.ComplexFormTypes.Add(complexFormType);
         patch.Remove(entry => entry.ComplexFormTypes, 0);
-        var changes = Entry.ChangesFromJsonPatch(_entry, patch);
+        var changes = _entry.ToChanges(patch);
         var removeComplexFormTypeChange = changes.Should().ContainSingle().Which.Should()
             .BeOfType<RemoveComplexFormTypeChange>().Subject;
         removeComplexFormTypeChange.EntityId.Should().Be(_entry.Id);
@@ -174,7 +172,7 @@ public class JsonPatchEntryRewriteTests
         _entry.ComplexFormTypes.Add(complexFormType);
         var newComplexFormType = new ComplexFormType() { Id = Guid.NewGuid(), Name = new MultiString() };
         patch.Replace(entry => entry.ComplexFormTypes, newComplexFormType, 0);
-        var changes = Entry.ChangesFromJsonPatch(_entry, patch);
+        var changes = _entry.ToChanges(patch);
         var replaceComplexFormTypeChange = changes.Should().ContainSingle().Which.Should()
         .BeOfType<ReplaceComplexFormTypeChange>().Subject;
         replaceComplexFormTypeChange.EntityId.Should().Be(_entry.Id);
